requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework>=3.0
pytz==2021.1
sqlparse==0.4.1
get-docker-secret<=2.0.0
mysqlclient<=3.0.0
